workers/llm-worker/requirements.txt
openai==1.54.4
httpx==0.27.2
python-dotenv==1.0.0
camunda-external-task-client-python3==4.3.0
